+++ packages/components/atoms/f-button/stories/LinkButton.stories.js
@@ -41,6 +41,7 @@ export const LinkButtonComponent = (args, { argTypes }) => ({
     </div>`
 });
 
+
 LinkButtonComponent.argTypes = SharedButtonArgTypes;
 
 LinkButtonComponent.args = {