+++ packages/components/atoms/f-button/CHANGELOG.md
@@ -6,10 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 v3.2.0
 ------------------------------
-*October 18, 2021*
+*October 22, 2021*
 
 ### Changed
-- Button border style definition moved to button types.
+- Button border style definition moved to button types. Introdused `background-color()` mixin to apply the same colour for the background and for the border colour.
 - Each button size vertical padding reduced by 1px to make button height 56px/48px/40px/32px instead of 58px/50px/42px/34px.
 
 ### Removed